Debug method exception handling
issue: the method returns the result of it's previous call.
When reaching a node that does not have a next, the method
would throw an exception for the current calculation,
and make the previous result unreachable

change: when there is no next, the next node is set to null.
An exception is thrown when attempting to move from null